[skip ci] reorganized them related files
